Remove unused pa11y tests
I have searched the histroy and I don't think these tests have ever worked or been run. They have always been missing the .pa11yci.js file that is needed for them to run

My hunch is that these tests and the code for them was copy and pasted from another component that did something similar in the inital set up.

Why I am removing them now is that they rely on chalk that is causing issues with the npm7 upgrade.