Port n-newsletter-signup to n-magnet and refactor demo app
This ports the newsletter signup component from n-newsletter-signup and drops
support for old tech used in that repo so that we didn't need to make it work in the demo app here.

In the process we've simplified lots of the code from n-newletter-signup to remove parts that were left over from the days when this component was also use on ft.com/newsletters.

Additionally the demo app was very broken and not usable for testing changes to code so to get this working we learn us much on existing standardised Toolkit implementations in other front end apps (especially copying from https://github.com/Financial-Times/next-user-facing-app-template)